add defectdojo session and product class
