--- aspmwarp/.env.template
@@ -1,2 +1,0 @@
-DEFECTDOJO_URL="http://localhost:8080"
-DEFECTDOJO_API_KEY="your key"
