--- aspmwarp/dtrack/dtrack.py
@@ -1,4 +1,0 @@
-import requests
-
-class DependencyTrack:
-    pass
