+++ pyproject.toml
@@ -1,6 +1,6 @@
 [project]
 name = "aspmate"
-version = "0.1.0"
+version = "0.0.1"
 description = "Add your description here"
 readme = "README.md"
 requires-python = ">=3.12"
